+++ javaee/rt/src/java/org/aion/avm/core/IExternalState.java
@@ -148,7 +148,7 @@ public interface IExternalState {
                        String method,
                        Object[] params,
                        BigInteger value,
-                       int stepLimit);
+                       long stepLimit);
 
     int getOption();
